pocs/people-remover/requirements.txt
flask==3.1.0
python-dotenv==1.0.1
requests==2.32.3
Pillow==11.1.0
